Adapt error handler
